refactor: small refactoring for the conversion from config file to arguments instead.
Refs: #9
